change dtype

+++ src/mcts/model/base.py
@@ -25,7 +25,7 @@ class MCTSModel(SaveableMultiModel, CustomDecorator, metaclass=ABCMeta):
         action_spec = game_spec.move_spec
         observation_spec = game_spec.observation_spec
 
-        self.action_range = np.stack((action_spec.minimum, action_spec.maximum), axis=0)
+        self.action_range = np.stack((action_spec.minimum, action_spec.maximum), axis=0, dtype=np.float32)
         self.action_shape = action_spec.shape
         self.observation_range = (
             np.stack((observation_spec.minimum, observation_spec.maximum), axis=0)
